wip: Add AUSD back to allow for calling fungibles' functions
We need the AUSD variant in order to call burn_from and mint_into, which
take care of handling the state updates across the different orml_tokens
storages. The alternative would be to create storage aliases for all of
the storages and translate over all and replicate the state changes that
those two mentioned functions perform but that sounds pretty dangerous
to me.
 So we add CurrencyId::AUSD with a deprecated flag, which we will drop
 once this migration gets executed.