Allow `estimate_gas` for contract creation
The `estimate_gas` RPC uses the same pathway as a `call` RPC, so
we need to allow contract creation in this path so that `estimate_gas`
of contract creation will work.

This does not affect any allow/deny-listing of actual contract creation,
since that filtering will happen on the transaction itself rather than
the `estimate_gas` call.