Fix block-rewards/tests/joining_leaving_collators
So the issue here is that in do_init_collator we would mint the
StakeAmount into the collator's account and then call
Rewards::deposit_stake which in turn calls Currency::can_hold, which
fails because the new version of fungibles/tokens etc doesn't support
ExistentialDeposit = 0, so that can_hold check fails since it can NOT
hold the entire balance (i.e, the entire balance of said account). So
the solution here involves adding ExistentialDeposit as a Config value
and, when minting the balance in do_init_collator, add the
ExistentialDeposit to it. By doing so, the tests also need to take the
ED value into account when checking values.